+++ attributes/swap.rb
@@ -18,5 +18,5 @@
 # limitations under the License.
 #
 
-default['rs-base']['swap']['size'] = 1 # GB
+default['rs-base']['swap']['size'] = 1024 # MB
 default['rs-base']['swap']['file'] = '/mnt/ephemeral/swapfile'